requirements.txt
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
google-api-core==2.11.0
google-auth==2.16.0
google-cloud-core==2.3.2
google-cloud-speech==2.18.0
google-cloud-storage==2.7.0
google-cloud-vision==3.3.1
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.1
grpcio-status==1.51.1
idna==3.4
Pillow==9.4.0
proto-plus==1.22.2
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.13
requests==2.28.2
rsa==4.9
six==1.16.0
SpeechRecognition==3.9.0
urllib3==1.26.14
